requirements.txt
Pillow==11.0.0
tqdm==4.66.5
PyMuPDF==1.27.1
img2pdf==0.5.1